+++ crates/nostr-relay-builder/src/error.rs
@@ -8,6 +8,7 @@ use std::{fmt, io};
 
 #[cfg(feature = "tor")]
 use async_wsocket::native::tor;
+use nostr_relay_pool::pool;
 
 /// Relay builder error
 #[derive(Debug)]
@@ -17,8 +18,12 @@ pub enum Error {
     /// Tor error
     #[cfg(feature = "tor")]
     Tor(tor::Error),
+    /// Relay pool error
+    RelayPool(pool::Error),
     /// Relay already running
     AlreadyRunning,
+    /// Premature exit
+    PrematureExit,
 }
 
 impl std::error::Error for Error {}
@@ -29,7 +34,9 @@ impl fmt::Display for Error {
             Self::IO(e) => write!(f, "{e}"),
             #[cfg(feature = "tor")]
             Self::Tor(e) => write!(f, "{e}"),
+            Self::RelayPool(e) => e.fmt(f),
             Self::AlreadyRunning => write!(f, "the relay is already running"),
+            Self::PrematureExit => write!(f, "premature exit"),
         }
     }
 }
@@ -44,5 +51,11 @@ impl From<io::Error> for Error {
 impl From<tor::Error> for Error {
     fn from(e: tor::Error) -> Self {
         Self::Tor(e)
+    }
+}
+
+impl From<pool::Error> for Error {
+    fn from(e: pool::Error) -> Self {
+        Self::RelayPool(e)
     }
 }