+++ crates/nostr-relay-builder/src/prelude.rs
@@ -8,12 +8,11 @@
 #![allow(ambiguous_glob_reexports)]
 #![doc(hidden)]
 
-// External crates
 pub use nostr::prelude::*;
 pub use nostr_database::prelude::*;
+pub use nostr_relay_pool::relay::SyncOptions;
 
 pub use crate::builder::{self, *};
 pub use crate::local::{self, *};
-// Internal modules
 pub use crate::mock::{self, *};
 pub use crate::*;